Stop menu double tap bug on mobile
On mobile, it would register one click as the "hover" to put the
strawberry on the link, then the second click as the actual "click" to
go to the link. By checking if the device supports hovering, we can
avoid this issue for mobile, so that only one click is requires.